+++ src/ip234.ts
@@ -1,4 +1,5 @@
-import { TestProxyResult, createAxiosInstance, ProxyConfig, TestProxyError, TestProxyErrorCode } from './common';
+import { TestProxyResult, TestProxyError, TestProxyErrorCode } from './common';
+import { createRequester, CreateRequesterOptions } from './requester';
 
 /**
  * IP234结果
@@ -42,13 +43,13 @@ interface Ip234Result {
 
 /**
  * 测试代理通过IP234
- * @param proxyConfig 代理配置或代理URL
+ * @param createRequesterOptions 创建请求器选项
  * @returns 代理测试结果
  */
-export async function testProxyInfoByIp234(proxyConfig?: ProxyConfig | string): Promise<TestProxyResult> {
-  const axios = createAxiosInstance(proxyConfig);
+export async function testProxyInfoByIp234(createRequesterOptions?: CreateRequesterOptions): Promise<TestProxyResult> {
+  const requester = createRequester(createRequesterOptions);
   const startTime = Date.now();
-  const { data } = await axios.get<Ip234Result>('https://ip234.in/ip.json');
+  const data = await requester.get<Ip234Result>('https://ip234.in/ip.json');
   const latency = Date.now() - startTime;
   if (!data) throw new TestProxyError('IP234 检测渠道返回结果为空', TestProxyErrorCode.DETECTION_CHANNEL_ERROR);
   return {